Final Assignment 3
Still having issues with the dropdown, some of the alignments, the
slide show, and the social media icons.  The icons were really
confusing and borrowing code from the main site helped sometimes but
not at others.   This was a fun challenge and it made me realize how
little I know about layout building.